Simplify GenericMod API, move to mods_base
