+++ src/models/mod_examples.py
@@ -2,7 +2,7 @@
 from typing import Dict, Sequence
 
 from models.abilities_base import Ability
-from models.ability_examples import FireLaser, Repair
+from models.ability_examples import FireLaser
 from models.mods_base import Mod
 from models.states import AttributeType, State, Attribute, Skill
 
@@ -76,23 +76,3 @@ class BasicLaser(Mod):
         return self._ability,
 
 
-class GenericMod(Mod):
-
-    def __init__(self, states_granted: Sequence[State] = (),
-                 attribute_modifiers: Dict[AttributeType, int] = None,
-                 abilities_granted: Sequence[Ability] = ()):
-        if attribute_modifiers is None:
-            attribute_modifiers = {}
-
-        self._states = states_granted
-        self._attr_mods = attribute_modifiers
-        self._abilities = abilities_granted
-
-    def states_granted(self) -> Sequence[State]:
-        return self._states
-
-    def attribute_modifiers(self) -> Dict[AttributeType, int]:
-        return self._attr_mods
-
-    def abilities_granted(self) -> Sequence[Ability]:
-        return self._abilities
