+++ src/models/mods_base.py
@@ -1,6 +1,6 @@
 """Base implementation of mods and inventory."""
 import abc
-from typing import Sequence, Dict
+from typing import Sequence, Dict, Union
 
 from models.abilities_base import Ability
 from models.states import State, AttributeType
@@ -20,3 +20,29 @@ class Mod(metaclass=abc.ABCMeta):
     @abc.abstractmethod
     def abilities_granted(self) -> Sequence[Ability]:
         """Abilities granted by this mod."""
+
+
+class GenericMod(Mod):
+
+    def __init__(self, states_granted: Union[State, Sequence[State]] = (),
+                 attribute_modifiers: Dict[AttributeType, int] = None,
+                 abilities_granted: Union[Ability, Sequence[Ability]] = ()):
+        if isinstance(states_granted, State):
+            states_granted = states_granted,
+        if attribute_modifiers is None:
+            attribute_modifiers = {}
+        if isinstance(abilities_granted, Ability):
+            abilities_granted = abilities_granted,
+
+        self._states = states_granted
+        self._attr_mods = attribute_modifiers
+        self._abilities = abilities_granted
+
+    def states_granted(self) -> Sequence[State]:
+        return self._states
+
+    def attribute_modifiers(self) -> Dict[AttributeType, int]:
+        return self._attr_mods
+
+    def abilities_granted(self) -> Sequence[Ability]:
+        return self._abilities
